requirements.txt
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.21.0
pandas==1.3.0
requests==2.25.1
urllib3==1.26.6
